[BACKLOG-14008] - Mavenize pentaho-vertica-bulkloader
