--- .classpath
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="src" path="test-src"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
-	<classpathentry kind="lib" path="/Kettle trunk/libswt/win32/swt.jar"/>
-	<classpathentry kind="lib" path="/Kettle trunk/libswt/jface.jar"/>
-	<classpathentry combineaccessrules="false" kind="src" path="/Kettle trunk"/>
-	<classpathentry kind="lib" path="libext/vertica-jdk5-6.0.0-0.jar"/>
-	<classpathentry kind="output" path="classes"/>
-</classpath>
