+++ assemblies/plugin/src/main/resources/version.xml
