+++ assemblies/pom.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>org.pentaho</groupId>
+    <artifactId>vertica-bulkloader</artifactId>
+  <version>7.1-SNAPSHOT</version>
+  </parent>
+  <artifactId>vertica-bulkloader-assemblies</artifactId>
+  <version>7.1-SNAPSHOT</version>
+  <packaging>pom</packaging>
+  <name>${project.artifactId}</name>
+  <description></description>
+  <modules>
+    <module>plugin</module>
+  </modules>
+</project>
