--- build.properties
@@ -1,7 +1,0 @@
-dependency.kettle.revision=7.1-SNAPSHOT
-project.revision=7.1-SNAPSHOT
-ivy.artifact.id=VerticaBulkLoader
-ivy.artifact.group=pentaho
-Implementation-Title=VerticaBulkLoader
-distrib=dist
-javadoc.packagenames=org.pentaho.di.*
