+++ core/src/main/java/org/pentaho/di/verticabulkload/VerticaBulkLoader.java
