+++ core/src/main/java/org/pentaho/di/verticabulkload/VerticaBulkLoaderData.java
