+++ core/src/main/java/org/pentaho/di/verticabulkload/nativebinary/ColumnSpec.java
