+++ core/src/main/resources/org/pentaho/di/verticabulkloader/messages/messages_de_DE.properties
