+++ core/src/main/resources/org/pentaho/di/verticabulkloader/messages/messages_fr_FR.properties
