+++ core/src/main/resources/org/pentaho/di/verticabulkloader/messages/messages_it_IT.properties
