+++ core/src/main/resources/vertica.png
