+++ core/src/test/java/org/pentaho/di/verticabulkload/VerticaBulkLoaderMetaInjectionTest.java
