+++ core/src/test/java/org/pentaho/di/verticabulkload/VerticaBulkLoaderMetaTest.java
