+++ core/src/test/java/org/pentaho/di/verticabulkload/nativebinary/ColumnSpecTimestampEncodingTest.java
