--- ivy.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ivy-module version="2.0" xmlns:m="http://ant.apache.org/ivy/maven">
-    <info organisation="${ivy.artifact.group}" module="${ivy.artifact.id}" revision="${project.revision}" />
-    
-    <configurations>
-    	<conf name="default"/>
-    	<conf name="provided"/>
-    	<conf name="runtime"/>
-    	<conf name="source" />
-        <conf name="zip" />
-    	<conf name="test" visibility="private"/>
-    </configurations>
-    
-    <publications>
-    	<artifact name="${ivy.artifact.id}" type="zip" conf="zip" />
-    	<artifact name="${ivy.artifact.id}" m:classifier="sources" type="source" ext="jar" conf="source" />
-    </publications>
-    
-    <dependencies defaultconf="default->default">
-        <dependency org="pentaho-kettle" name="kettle-core"   rev="${dependency.kettle.revision}" changing="true" conf="provided->default"/>
-        <dependency org="pentaho-kettle" name="kettle-engine" rev="${dependency.kettle.revision}" changing="true" conf="provided->default"/>
-        <dependency org="pentaho-kettle" name="kettle-ui-swt"     rev="${dependency.kettle.revision}" changing="true" conf="provided->default"/>        
-
-        <!--  Testing dependencies -->
-        <dependency org="pentaho-kettle" name="kettle-engine-test" rev="${dependency.kettle.revision}" changing="true" conf="test->default"/>
-        <dependency org="junit"          name="junit"              rev="4.4" conf="test->default" />
-        <dependency org="org.mockito"    name="mockito-all"        rev="1.9.5" transitive="false" conf="test->default"/>
-    </dependencies>
-</ivy-module>
