--- package-ivy.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<ivy-module version="2.0" xmlns:m="http://ant.apache.org/ivy/maven">
-  <info organisation="${ivy.artifact.group}" module="${ivy.artifact.id}" revision="${project.revision}" />
-  <configurations>
-    <conf name="default"/>
-        <conf name="runtime"/>
-        <conf name="source" />
-        <conf name="zip" />
-        <conf name="test" visibility="private"/>
-  </configurations>
-
-  <publications>
-    <artifact name="${ivy.artifact.id}" type="zip" conf="zip"/>
-  </publications>
-
-  <dependencies defaultconf="default->default">
-  </dependencies>
-</ivy-module>
